객체 지향적으로 변경
